# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Afro Fish platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid admin PIN
  /admin/players:
    get:
      operationId: getAdminPlayers
      tags: [admin]
      summary: List all players
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [admin]
      summary: Create a new player
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /admin/reset-pin:
    post:
      operationId: resetPlayerPin
      tags: [admin]
      summary: Reset a player's PIN
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
        - name: playerId
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPinInput"
      responses:
        "200":
          description: PIN updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /admin/delete-player:
    post:
      operationId: deletePlayer
      tags: [admin]
      summary: Delete a player
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
        - name: playerId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /admin/reload-credits:
    post:
      operationId: reloadPlayerCredits
      tags: [admin]
      summary: Reload credits (cash in) to a player account
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
        - name: playerId
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditReloadInput"
      responses:
        "200":
          description: Updated player with bonus info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReloadResult"
  /admin/cashout:
    post:
      operationId: cashoutPlayer
      tags: [admin]
      summary: Process cash out for a player
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
        - name: playerId
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CashoutInput"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /admin/rescue:
    post:
      operationId: rescuePlayer
      tags: [admin]
      summary: Give comeback credits to a low-balance player
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
        - name: playerId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /admin/transactions:
    get:
      operationId: getAdminTransactions
      tags: [admin]
      summary: Get all transactions
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: playerId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Get traffic and revenue analytics
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
        - name: days
          in: query
          required: false
          schema:
            type: integer
            enum: [1, 7, 30, 90]
            default: 30
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnalytics"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get platform settings
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    post:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update platform settings
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform stats
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/game-config:
    get:
      operationId: getGameConfig
      tags: [admin]
      summary: Get win rate configs for all games and tiers
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Game config list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameConfigList"
  /admin/set-game-config:
    post:
      operationId: updateGameConfig
      tags: [admin]
      summary: Update win rate for a game and tier
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
        - name: game
          in: query
          required: true
          schema:
            type: string
        - name: tier
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameConfigInput"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameConfig"
  /admin/bonus-config:
    get:
      operationId: getBonusConfig
      tags: [admin]
      summary: Get bonus configuration
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bonus config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonusConfig"
    post:
      operationId: updateBonusConfig
      tags: [admin]
      summary: Update bonus configuration
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BonusConfigInput"
      responses:
        "200":
          description: Updated bonus config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonusConfig"
components:
  schemas:
...
    AdminLoginInput:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
    PlayerInput:
      type: object
      required: [name, pin]
      properties:
        name:
          type: string
        pin:
          type: string
        initialCredits:
          type: number
          default: 0
    ResetPinInput:
      type: object
      required: [newPin]
      properties:
        newPin:
          type: string
    CreditReloadInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        note:
          type: string
    CashoutInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        note:
          type: string
...
    ReloadResult:
      type: object
      required: [player, bonusApplied, bonusType]
      properties:
        player:
          $ref: "#/components/schemas/Player"
        bonusApplied:
          type: integer
        bonusType:
          type: string
    GameConfigInput:
      type: object
      required: [winRate]
      properties:
        winRate:
          type: number
          minimum: 0
          maximum: 100
...
    Player:
      type: object
      required: [id, name, balance, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        balance:
          type: number
        totalWon:
          type: number
        totalLost:
          type: number
        gamesPlayed:
          type: integer
        totalKills:
          type: integer
        killTrophiesClaimed:
          type: integer
        lastLoginDate:
          type: ["string", "null"]
        firstDepositDone:
          type: boolean
        createdAt:
          type: string
...
    AdminSession:
      type: object
      required: [adminKey]
      properties:
        adminKey:
          type: string
    Transaction:
      type: object
      required: [id, playerId, playerName, type, amount, createdAt]
      properties:
        id:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
        type:
          type: string
          enum: [credit_in, credit_out, game_win, game_loss, bonus]
        amount:
          type: number
        note:
          type: ["string", "null"]
        game:
          type: ["string", "null"]
        tier:
          type: ["string", "null"]
        createdAt:
          type: string
...